Add Tree.add_and_reg() and Tree.remove_and_unreg()
Instead of using add() to add each object path and then registering all of
them with set_registered(), these two functions do both. The key benefit
is that it is now possible to remove and unregister.

Clearly, mixing the two new methods with the two existing ones is not
a good idea. I'm not sure if it's better to programatically handle this,
or just make it clear in the docs to not mix them

Signed-off-by: Andy Grover <agrover@redhat.com>