Fix polling failure due to read_write_dispatch()
`ADriver` falls back to `iter(0)` when watched
socket is not ready.

`iter(0)` calls `dbus_connection_read_write_dispatch()`.

If the data arrived after the socket check but before
`dbus_connection_read_write_dispatch()`, then the latter
would empty the socket but wouldn't yield the message
to `filter_message_cb()` until the next invocation of itself.

But the next invocation doesn't happen, due to to
`Nothing => return` in `handle_items()`.

So neither `ADriver` is polled again, because the socket
is already empty, nor `read_write_dispatch()` is called
by anything again.

And thus `aconnection_test()` hangs indefinitely.

`iter(0)` with `read_write_dispatch()` is not needed,
and removing it and relying on `watch_handle()` and
`iter(None)` with pure `dispatch()` fixes the bug.