Made TryFrom<&MessageItem> impl's use peel()
In general, callers of inner() are looking for actual value inside a
MessageItem, even if it's nested inside a MessageItem::Variant. By adding peel()
to the TryFrom impls, that behavior is acheived.

For applications where it is important to distinguish between a Variant and the
raw MessageItem, match and if let constructs still work unchanged.

Note that this is a minor breaking change: try_from() (and by extension, inner())
will now sometimes return Ok(_) values when they previously returned Err(_).