dbus: guarantee representations of arrays of 'simple' types.
Sometimes a large array might be stored behind a variant; for example,
the 'show a notification' message in the desktop notifications spec can
store an image in a dict of variants. So we should guarantee that you
can cast that variant to a Vec of the right type.

We don't make that guarantee for more complicated types like structs or
dicts because their representations are more up-in-the-air.