Do not include the ManagedObjects implementor in the result
According to the spec, this object is supposed to be omitted from the
result of the GetManagedObjects() call.

The spec link:
https://dbus.freedesktop.org/doc/dbus-specification.html#standard-interfaces-objectmanager.

The text:
The return value of this method is a dict whose keys are object paths.
All returned object paths are children of the object path implementing
this interface, i.e. their object paths start with the ObjectManager's
object path plus '/'.

Split an existing test into two tests, so that the result of a
GetManagedObjects() call can be handled separately from tests about
properties.