Dbus: Allow 25 function arguments instead of 13
This limit was increased in 389d41066e2dddfd62719282519121ff45b9f3d2,
however this is still insufficient for some methods (ie. org.gpsd's fix
signal).

Obviously this is just pushing an arbitrary limit further back, but I
figured running out of letters in the alphabet was a good place to stop
for now. Bonus points for finding the letter that couldn't be used.